+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/helpers/DBHelper.kt
@@ -124,7 +124,7 @@ class DBHelper private constructor(val context: Context) : SQLiteOpenHelper(cont
         val id = mDb.insert(MAIN_TABLE_NAME, null, eventValues)
         event.id = id.toInt()
 
-        if (event.repeatInterval != 0 && event.parentId == 0) {
+        if (event.repeatInterval != 0 && event.parentId == 0L) {
             val metaValues = fillMetaValues(event)
             mDb.insert(META_TABLE_NAME, null, metaValues)
         }
@@ -151,7 +151,7 @@ class DBHelper private constructor(val context: Context) : SQLiteOpenHelper(cont
                 val id = mDb.insert(MAIN_TABLE_NAME, null, eventValues)
                 event.id = id.toInt()
 
-                if (event.repeatInterval != 0 && event.parentId == 0) {
+                if (event.repeatInterval != 0 && event.parentId == 0L) {
                     val metaValues = fillMetaValues(event)
                     mDb.insert(META_TABLE_NAME, null, metaValues)
                 }
@@ -271,7 +271,7 @@ class DBHelper private constructor(val context: Context) : SQLiteOpenHelper(cont
 
         childEvent.apply {
             id = 0
-            parentId = parentEventId
+            parentId = parentEventId.toLong()
             startTS = 0
             endTS = 0
             if (childImportId != null) {
